gpu region cordon barrows crypts
shade catacombs "must" be included because the barrows crypts and dungeon are on different planes -> kind of moot since the immersion is broken anyway.

on the other hand, if only cordoning the region for the crypt/dungeon then scene reload must be forced, or else entities and transforming objects will appear in the void (e.g. npcs, doors opening/closing) -> not good.